requirements.txt
# API and data handling
requests==2.31.0
pandas==2.2.0
numpy>=2.0.0
python-dotenv==1.0.0

# Data storage
pyarrow==14.0.1  # For parquet files
psycopg2-binary==2.9.9  # For PostgreSQL

# Scheduling
schedule==1.2.1

# Data analysis
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0

# Development tools
jupyter==1.0.0
black==23.9.1
flake8==6.1.0
pytest==7.4.2